+++ test/NonEmptyArray.ts
@@ -381,8 +381,7 @@ describe('NonEmptyArray', () => {
     U.deepStrictEqual(Sh.show(['a', 'b', 'c']), `["a", "b", "c"]`)
   })
 
-  it('alt / concat', () => {
-    U.deepStrictEqual(_.concat(['a'], []), ['a'])
+  it('alt', () => {
     U.deepStrictEqual(
       pipe(
         ['a'],
@@ -549,5 +548,17 @@ describe('NonEmptyArray', () => {
 
   it('concatW', () => {
     U.deepStrictEqual(pipe(['a'], _.concatW(['b'])), ['a', 'b'])
+  })
+
+  it('concat', () => {
+    U.deepStrictEqual(pipe(['a'], _.concat(['b'])), ['a', 'b'])
+    U.deepStrictEqual(pipe([], _.concat(['b'])), ['b'])
+    U.deepStrictEqual(pipe(['a'], _.concat<string>([])), ['a'])
+    // tslint:disable-next-line: deprecation
+    U.deepStrictEqual(_.concat(['a'], ['b']), ['a', 'b'])
+    // tslint:disable-next-line: deprecation
+    U.deepStrictEqual(_.concat(['a'], []), ['a'])
+    // tslint:disable-next-line: deprecation
+    U.deepStrictEqual(_.concat([], ['b']), ['b'])
   })
 })